% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder_wrap.R
\name{adder_wrap}
\alias{adder_wrap}
\title{Wrapped python including import
This imports a python module internally, uses it, and returns the output}
\usage{
adder_wrap(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
Wrapped python including import
This imports a python module internally, uses it, and returns the output
}
\examples{
adder_wrap(2,3)
}
