% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsummary_wrap.R
\name{pdsummary_wrap}
\alias{pdsummary_wrap}
\title{simple wrapper of a pandas summarise}
\usage{
pdsummary_wrap(df, group)
}
\arguments{
\item{df}{a dataframe}

\item{group}{character, grouping column}
}
\description{
simple wrapper of a pandas summarise
}
