% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipython_wrap.R
\name{wrap_select}
\alias{wrap_select}
\title{wrap pandas column select}
\usage{
wrap_select(...)
}
\arguments{
\item{...}{df, columns to select as character vector}
}
\description{
wrap pandas column select
}
